requirements.txt
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Validation and Settings
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
redis>=5.0.0
alembic>=1.12.0

# LLM Providers
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
langchain-core>=0.1.0

# Monitoring and Logging
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
sentry-sdk>=1.32.0

# Development Tools
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0
